Add examples directory, fixes to analog_temp library, details fixed in main.h, deleted unusued can_test.c
