+++ examples/analog_temp_read/makefile
@@ -1,0 +1,4 @@
+PROG = analog_temp_read
+# SRC should only include necessary files
+SRC = $(addprefix ../../src/,analog_temp.c)
+include ../makefile
